Create enemy types
Two new enemy prefabs have been created: Berzerker and Runner.
Both have their own AI classes, derived from the script in the previous commit.
