+++ Assets/Prefabs/Characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 22204ca3057cb2d45af66ba861d805a5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
