+++ Assets/Prefabs/Characters/Berzerker.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bb7a86a2b3efe3241afba503320ea4f9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
