+++ Assets/Prefabs/Characters/Runner.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 15c7c4d44912bc44599ea22b862d6de9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
