+++ Assets/Scripts/AI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d0019c3be3a508f4a974208b20587ead
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
