+++ Assets/Scripts/AI/AIBerzerker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dde457bfe3bdabb4798fb8059bf1d3b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
